+++ commands/validator/utils.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-package importer
+package validator
 
 import (
 	"encoding/json"
